+++ bin/copy-lforms.js
@@ -17,7 +17,7 @@ const path = require('path');
 let downloadUrl = 'https://clinicaltables.nlm.nih.gov/lforms-versions';
 
 // Change version here to update LForms package.
-const version = '33.0.0';
+const version = '33.4.1';
 
 // External scripts could change version?
 if(version?.length) {