+++ cypress/e2e/item-level/item-level.cy.js
@@ -359,9 +359,52 @@ describe('Home page', () => {
       cy.uploadFile(sampleFile, true);
       cy.get('#title').should('have.value', 'Answer options form');
       cy.contains('button', 'Edit questions').click();
+      cy.get('lfb-answer-option table > tbody > tr:nth-of-type(1)').as('firstOption');
+      cy.get('lfb-answer-option table > tbody > tr:nth-of-type(2)').as('secondOption');
+
+      cy.get('@firstOption').find('td:nth-child(1) input').should('have.value', 'd1');
+      cy.get('@firstOption').find('td:nth-child(2) input').should('have.value', 'a');
+      cy.get('@firstOption').find('td:nth-child(3) input').should('have.value', 's');
+      cy.get('@firstOption').find('td:nth-child(4) input').should('have.value', '1');
+      cy.get('@firstOption').find('td:nth-child(5) input').should('be.visible').and('not.be.checked');
+      cy.get('@secondOption').find('td:nth-child(1) input').should('have.value', 'd2');
+      cy.get('@secondOption').find('td:nth-child(2) input').should('have.value', 'b');
+      cy.get('@secondOption').find('td:nth-child(3) input').should('have.value', 's');
+      cy.get('@secondOption').find('td:nth-child(4) input').as('secondScore')
+        .should('have.value', '2');
+      cy.get('@secondOption').find('td:nth-child(5) input').as('secondDefaultRadio')
+        .should('be.visible').and('be.checked');
+
       cy.questionnaireJSON().should((qJson) => {
         expect(qJson.item[0].answerOption).to.deep.equal(fixtureJson.item[0].answerOption);
-        expect(qJson.item[0].initial).to.deep.equal(fixtureJson.item[0].initial);
+      });
+
+      cy.get('@secondScore').clear().type('22');
+      cy.get('lfb-answer-option table+button').click();
+      cy.get('lfb-answer-option table > tbody > tr:nth-of-type(3)').as('thirdOption').should('be.visible');
+      cy.get('@thirdOption').find('td:nth-child(1) input').type('d3');
+      cy.get('@thirdOption').find('td:nth-child(2) input').type('c');
+      cy.get('@thirdOption').find('td:nth-child(3) input').type('s');
+      cy.get('@thirdOption').find('td:nth-child(4) input').type('33');
+      cy.get('@thirdOption').find('td:nth-child(5) input').as('thirdDefaultRadio').click();
+
+      const ORDINAL_URI = 'http://hl7.org/fhir/StructureDefinition/ordinalValue';
+      cy.questionnaireJSON().should((qJson) => {
+        expect(qJson.item[0].answerOption).to.deep.equal([
+          {
+            valueCoding: {display: 'd1', code: 'a', system: 's'},
+            extension: [{url: ORDINAL_URI, valueDecimal: 1}]
+          },
+          {
+            valueCoding: {display: 'd2', code: 'b', system: 's'},
+            extension: [{url: ORDINAL_URI, valueDecimal: 22}]
+          },
+          {
+            valueCoding: {display: 'd3', code: 'c', system: 's'},
+            extension: [{url: ORDINAL_URI, valueDecimal: 33}],
+            initialSelected: true
+          },
+       ]);
       });
     });
 
@@ -853,6 +896,10 @@ describe('Home page', () => {
       });
 
       cy.get('@unit0').clear().type('xxxx').blur().should('have.value', 'xxxx');
+
+      // The blur() event may not be enough to update the form. Use some UI events to trigger the update.
+      cy.contains('button', 'Preview').click();
+      cy.contains('mat-dialog-actions button', 'Close').click();
 
       cy.questionnaireJSON().should((qJson) => {
         expect(qJson.item[0].initial[0]).to.deep.equal({