+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 This project follows [Semantic Versioning](http://semver.org/).
 
+## [9.1.3] 2023-07-20
+### Fixed
+- Fixed a bug which was missing scores in answerOption field.
+- Replaced initial[x] objects with answerOption[x].initialSelected flag to indicate
+  default answers in choice/open-choice type items.
+- Updated lforms to 33.4.1 to fix a bug with list positioning of autocomplete input box
+  in preview dialog.
+
 ## [9.1.2] 2023-07-11
 ### Added
 - Add item control extension. For now, the support restricted to autocomplete in ValueSet fields.